refactor(app): give the tab's panes and layout persistence their own siblings
`tab.rs` carried 1,661 production lines. Two groups move out unchanged:
`tab/panes.rs` takes the tab's pane addressing -- the context column's
collapse and order, the index and side accessors, focus, and the tape
the flow pane shows -- and `tab/layout.rs` takes what the tab persists
about its panes: the canvas layout switch, the pending layout it
applies, the layout a restored pane opens with, the canvas restore and
the bar-spec changes a pane applies.

Pure moves: every body is unchanged and no visibility needed widening,
because every moved method was already `pub` or `pub(crate)`. The feed
drain is not touched -- it lives in `tab/feed.rs` and stays there. What
stays in `tab.rs` is the tab's own construction and feed attachment, the
chip label, the capabilities it reports and the latency it publishes;
the `QUANTICK_PANE_COLLAPSED` declaration stays beside its read in
`Tab::new`.
