+++ crates/app/src/tab.rs
@@ -46,6 +46,8 @@ use std::path::PathBuf;
 mod canvas;
 mod feed;
 mod history;
+mod layout;
+mod panes;
 mod strategies;
 
 pub use canvas::CanvasChrome;
@@ -556,114 +558,6 @@ pub struct Tab {
 }
 
 impl Tab {
-    /// Put the context column away, or bring it back.
-    ///
-    /// **The one collapse path.** The divider drag, the rail, the View menu,
-    /// `Ctrl+0` and `layout.pane.collapse` all arrive here. Before this
-    /// existed the only way a *trader* could collapse the column was a mouse
-    /// drag, while an assistant had a named call for it — the second operator
-    /// holding a capability the first could not reach from the keyboard,
-    /// which is the rule inverted.
-    ///
-    /// `split_fraction` is deliberately untouched: it is the width the column
-    /// springs back to, and spending it here would hand back a different chart
-    /// from the one that was put away.
-    ///
-    /// Returns whether anything changed.
-    pub fn set_context_collapsed(&mut self, collapsed: bool) -> bool {
-        if self.context_collapsed == collapsed {
-            return false;
-        }
-        self.context_collapsed = collapsed;
-        true
-    }
-
-    /// Move the context pane at `from` to `to`, keeping the rest in order.
-    ///
-    /// **The one reposition path.** The View menu, the keyboard and the
-    /// control plane all arrive here, so none of them can grow its own idea of
-    /// what moving a pane does; a drag gesture, when it lands, is sugar over
-    /// this call rather than a second implementation of it.
-    ///
-    /// Addresses are [`PaneIndex`]es, so `0` names the flow pane. The flow
-    /// pane does not move: it is the protagonist and its column is the one
-    /// thing every preset agrees on. Refused rather than clamped — a caller
-    /// that asked to move the heatmap meant something this cannot do, and
-    /// quietly moving a different pane would be worse than saying no.
-    ///
-    /// Returns whether anything moved.
-    pub fn move_context_pane(&mut self, from: PaneIndex, to: PaneIndex) -> bool {
-        let (Some(from_slot), Some(to_slot)) = (from.checked_sub(1), to.checked_sub(1)) else {
-            return false;
-        };
-        if from_slot >= self.time_panes.len() || to_slot >= self.time_panes.len() {
-            return false;
-        }
-        if from_slot == to_slot {
-            return false;
-        }
-        let pane = self.time_panes.remove(from_slot);
-        self.time_panes.insert(to_slot, pane);
-        // Focus follows the pane the trader just moved, so the next command
-        // lands on the chart they were working with rather than on whichever
-        // one slid into its place.
-        self.focus = PaneSide::Time(to_slot);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "LAYOUT_PANE_MOVED",
-            tab_id = self.id,
-            from = from,
-            to = to,
-            "a context pane was moved within the stack"
-        );
-        true
-    }
-
-    /// How many panes this tab holds, drawn or not.
-    #[must_use]
-    pub fn pane_count(&self) -> PaneIndex {
-        1 + self.time_panes.len()
-    }
-
-    /// The pane at `index`: `0` is the flow pane, `1..` the context stack.
-    #[must_use]
-    pub fn pane_at(&self, index: PaneIndex) -> Option<&ChartPane> {
-        match index {
-            0 => Some(&self.flow_pane),
-            other => self.time_panes.get(other - 1),
-        }
-    }
-
-    /// The pane at `index`, mutably.
-    pub fn pane_at_mut(&mut self, index: PaneIndex) -> Option<&mut ChartPane> {
-        match index {
-            0 => Some(&mut self.flow_pane),
-            other => self.time_panes.get_mut(other - 1),
-        }
-    }
-
-    /// The first context pane, if this tab has built one.
-    ///
-    /// Most of the chrome speaks about *the* context chart because most
-    /// layouts show one. The ones that show more reach for `time_panes`
-    /// directly; this is the convenience, not the truth.
-    #[must_use]
-    pub fn time_pane(&self) -> Option<&ChartPane> {
-        self.time_panes.first()
-    }
-
-    /// The first context pane, mutably.
-    pub fn time_pane_mut(&mut self) -> Option<&mut ChartPane> {
-        self.time_panes.first_mut()
-    }
-
-    /// Whether this tab has built any context pane at all.
-    #[must_use]
-    pub fn has_time_pane(&self) -> bool {
-        !self.time_panes.is_empty()
-    }
-
     /// A tab on `feed_id`/`symbol`, already streaming through `feed`, showing
     /// bar `spec`.
     ///
@@ -903,464 +797,6 @@ impl Tab {
         drop(tx);
     }
 
-    /// Whether this tab has trouble worth showing on its chip while it sits in
-    /// the background (§11: honesty at a glance).
-    ///
-    /// A recording never reports transport trouble — it has no transport — and
-    /// "still connecting" is not trouble either. This is a feed that had a
-    /// connection and lost it, or one asking the user to fix something.
-    #[must_use]
-    pub fn needs_attention(&self) -> bool {
-        self.replay.is_none()
-            && (self.feed_connection == FeedConnectionState::Reconnecting
-                || matches!(self.notice, FeedNotice::Attention { .. }))
-    }
-
-    /// The canvas width the last drawn frame used, for callers that have to
-    /// reason in pixels without a frame in hand — the control plane's resize,
-    /// which must honour the same floor a drag does.
-    ///
-    /// Zero before the first frame, which callers read as "no opinion" rather
-    /// than as a canvas of no width.
-    #[must_use]
-    pub fn last_canvas_width(&self) -> f32 {
-        self.last_canvas_width
-    }
-
-    /// Whether any context chart is actually on screen.
-    ///
-    /// The layout has to *hold* one, the tab has to have *built* one, and the
-    /// column must not be collapsed. Three conditions that chrome kept
-    /// re-deriving one variant at a time — and got wrong twice: the legend
-    /// gate matched `TimeAndFlow` alone, so the stacked charts drew no legend
-    /// and a collapsed column drew one over the flow chart against a stale
-    /// rect.
-    #[must_use]
-    pub fn shows_context_charts(&self) -> bool {
-        self.layout.shows_time() && self.has_time_pane() && !self.context_collapsed
-    }
-
-    /// The pane the chrome speaks for. Only a split canvas has a choice to
-    /// make: a single-pane layout *is* its one visible pane, whatever the
-    /// last split left `focus` set to. Time falls back to the flow pane for
-    /// the frame between asking for the layout and the pane being built.
-    pub fn focused_side(&self) -> PaneSide {
-        // Read from what the layout *holds*, never from which variant it is.
-        // Matching one variant is how the three-pane canvas shipped with dead
-        // focus: a click set `self.focus` and every reader threw it away,
-        // because `TimeTimeAndFlow` was not in the arm.
-        // A collapsed column is not on screen, and focus on a pane nobody can
-        // see is worse than useless: `paper_hud_here` goes false for the one
-        // pane that *is* drawn, so order entry, the ladder and the trade HUD
-        // all go dead on the heatmap until the trader expands again.
-        if !self.has_time_pane() || !self.layout.shows_time() {
-            return PaneSide::Flow;
-        }
-        if !self.layout.shows_flow() {
-            // A layout with no flow pane has nothing to fall back *to*: the
-            // context chart is the only pane drawn, and the collapse flag
-            // names a column this layout does not carve. Read before the flag,
-            // or `Ctrl+0` on the Timeframe layout sent focus to a pane nobody
-            // draws — order entry, the ladder and the trade HUD all dead on
-            // the one chart on screen, with no rail to click to undo it.
-            return PaneSide::Time(0);
-        }
-        if self.context_collapsed {
-            return PaneSide::Flow;
-        }
-        // A slot the stack no longer shows — the focused pane was the bottom
-        // of a three-pane layout and the trader switched to two — falls back
-        // to the top context chart, never to a pane that is not drawn.
-        match self.focus {
-            PaneSide::Time(slot) if slot >= self.context_panes_shown() => PaneSide::Time(0),
-            focus => focus,
-        }
-    }
-
-    /// How many context panes the layout draws — bounded by how many exist,
-    /// for the frame between asking for a layout and its panes being built.
-    pub fn context_panes_shown(&self) -> usize {
-        self.layout
-            .kinds()
-            .iter()
-            .filter(|kind| **kind == PaneKind::Time)
-            .count()
-            .min(self.time_panes.len())
-    }
-
-    /// The pane on `side`, falling back to the flow pane when the time pane
-    /// has never been opened.
-    pub fn pane(&self, side: PaneSide) -> &ChartPane {
-        match side {
-            PaneSide::Time(slot) => self.time_panes.get(slot).unwrap_or(&self.flow_pane),
-            PaneSide::Flow => &self.flow_pane,
-        }
-    }
-
-    /// Whether a LAYERS lamp reads as on, and what blocks it if anything.
-    ///
-    /// One reading, two callers: the toolbar model built each frame and the
-    /// semantic scene an operator captures on demand. A lamp that told the
-    /// trader one thing and an assistant another would be worse than no scene
-    /// at all, so neither side gets its own copy of the question — both come
-    /// here, and here asks [`ChartPane::layer_switched_on`] and
-    /// [`ChartPane::layer_blocked`], which already resolve every layer to the
-    /// one field and the one gate that own it.
-    ///
-    /// The reading is the *switch*, not what the source lets through it: a
-    /// lamp lit from `layer_visible` reads dark while book capture is starting
-    /// and forever on a source with no book, so the trader presses an unlit
-    /// button and switches the layer they wanted off. The block beside it is
-    /// what says the source cannot fill it.
-    pub(crate) fn layer_toggle_state(
-        &self,
-        layer: ChartLayer,
-        style: &ChartStyle,
-        capabilities: FeedCapabilities,
-    ) -> (bool, Option<LayerBlock>) {
-        let pane = self.pane(self.layer_toggle_side(layer));
-        (
-            pane.layer_switched_on(layer, style),
-            pane.layer_blocked(layer, capabilities),
-        )
-    }
-
-    /// Which pane a LAYERS button speaks for.
-    ///
-    /// The footprint folds the pane's *own* retained trades, so its lamp
-    /// answers for the pane with focus: one lit from the flow pane while the
-    /// time pane has focus would report a layer the trader is not looking at.
-    /// The other three read the tape, and only the flow pane has one.
-    fn layer_toggle_side(&self, layer: ChartLayer) -> PaneSide {
-        match layer {
-            // Read off the tape, and only the flow pane has one. A time pane
-            // asked about these answers for machinery it does not own, which
-            // is what `ChartPane::layer_blocked` says in words.
-            ChartLayer::TapeChart
-            | ChartLayer::TapeHeatmap
-            | ChartLayer::TapeBubbles
-            | ChartLayer::Heatmap
-            | ChartLayer::Bubbles
-            | ChartLayer::LiveStrip
-            | ChartLayer::LaneMarks
-            | ChartLayer::FlowLegend
-            | ChartLayer::BookStatus
-            | ChartLayer::DepthGaps => PaneSide::Flow,
-            // The pane's own: the footprint folds the pane's retained trades,
-            // the rest are that canvas's chrome and its objects. A lamp lit
-            // from the flow pane while the time pane has focus would report a
-            // layer the trader is not looking at.
-            ChartLayer::Footprint
-            | ChartLayer::Grid
-            | ChartLayer::LastPrice
-            | ChartLayer::BackfillDivider
-            | ChartLayer::SeamDivider
-            | ChartLayer::Crosshair
-            | ChartLayer::PointerPrice
-            | ChartLayer::PointerTime
-            | ChartLayer::PaperTrading
-            | ChartLayer::TradePaint
-            | ChartLayer::Drawings => self.focused_side(),
-        }
-    }
-
-    /// See [`Self::pane`].
-    /// Point exactly one of this tab's panes at the object whose content is
-    /// being typed off-canvas, and clear the other.
-    ///
-    /// Both are written every time on purpose: the flag suppresses an
-    /// object's own painting, so a pane left holding a stale index keeps a
-    /// note invisible for the rest of the session. The tab is what knows
-    /// whether a time pane exists at all, which is why the loop lives here
-    /// rather than in the host.
-    pub fn set_content_editing(&mut self, target: Option<(PaneSide, usize)>) {
-        self.flow_pane.gestures.content_editing = target
-            .filter(|(side, _)| *side == PaneSide::Flow)
-            .map(|(_, index)| index);
-        for (slot, time) in self.time_panes.iter_mut().enumerate() {
-            time.gestures.content_editing = target
-                .filter(|(side, _)| *side == PaneSide::Time(slot))
-                .map(|(_, index)| index);
-        }
-    }
-
-    pub fn pane_mut(&mut self, side: PaneSide) -> &mut ChartPane {
-        match side {
-            PaneSide::Time(slot) => self.time_panes.get_mut(slot).unwrap_or(&mut self.flow_pane),
-            PaneSide::Flow => &mut self.flow_pane,
-        }
-    }
-
-    /// Every side this tab can address, flow first — the order
-    /// [`Self::panes`] walks. The one place "which panes exist" is spelled
-    /// out for the chrome, so a surface that asks each pane a question walks
-    /// the stack rather than the two sides the split used to have.
-    pub fn sides(&self) -> impl Iterator<Item = PaneSide> + '_ {
-        (0..self.pane_count()).map(PaneSide::from_index)
-    }
-
-    /// The pane every chrome surface reads from — see [`Self::focused_side`].
-    pub fn focused_pane(&self) -> &ChartPane {
-        self.pane(self.focused_side())
-    }
-
-    /// See [`Self::focused_pane`].
-    pub fn focused_pane_mut(&mut self) -> &mut ChartPane {
-        self.pane_mut(self.focused_side())
-    }
-
-    /// The pane holding the drawing selection — which is not always the
-    /// focused one.
-    ///
-    /// A shared mark can be taken from either chart it appears on, and it
-    /// stays in the store of the pane it was drawn on. So the inspector, the
-    /// keyboard and the object manager follow the *object*, not the pane the
-    /// pointer happens to be over: selecting a level on the time pane and
-    /// pressing Delete has to delete that level, wherever it lives.
-    ///
-    /// Exactly one pane holds a selection at a time
-    /// ([`Self::apply_shared_interactions`] drops the other's), so this asks
-    /// the focused pane first and takes the answer it finds.
-    pub fn drawing_side(&self) -> PaneSide {
-        let focused = self.focused_side();
-        if self.pane(focused).drawings.selected().is_some() || self.time_panes.is_empty() {
-            return focused;
-        }
-        self.sides()
-            .find(|side| *side != focused && self.pane(*side).drawings.selected().is_some())
-            .unwrap_or(focused)
-    }
-
-    /// The pane every drawing surface reads from — see [`Self::drawing_side`].
-    pub fn drawing_pane(&self) -> &ChartPane {
-        self.pane(self.drawing_side())
-    }
-
-    /// See [`Self::drawing_pane`].
-    pub fn drawing_pane_mut(&mut self) -> &mut ChartPane {
-        self.pane_mut(self.drawing_side())
-    }
-
-    /// Every pane holding this market's bars, on screen or not, each beside
-    /// the side it answers to — flow first, so a walk of one tab is stable and
-    /// a capture stays diffable against the one before it.
-    ///
-    /// An iterator and not a `Vec`: this is walked per frame by the control
-    /// plane's journal comparison, which must not touch the allocator on a
-    /// quiet frame.
-    pub fn panes(&self) -> impl Iterator<Item = (&ChartPane, PaneSide)> {
-        std::iter::once((&self.flow_pane, PaneSide::Flow)).chain(
-            self.time_panes
-                .iter()
-                .enumerate()
-                .map(|(slot, time)| (time, PaneSide::Time(slot))),
-        )
-    }
-
-    /// Every pane holding this market's bars, on screen or not. One tape, and
-    /// however many charts the layout has ever shown read off it.
-    pub fn panes_with_sides_mut(&mut self) -> impl Iterator<Item = (&mut ChartPane, PaneSide)> {
-        let Self {
-            flow_pane,
-            time_panes,
-            ..
-        } = self;
-        std::iter::once((flow_pane, PaneSide::Flow)).chain(
-            time_panes
-                .iter_mut()
-                .enumerate()
-                .map(|(slot, pane)| (pane, PaneSide::Time(slot))),
-        )
-    }
-
-    /// See [`Self::panes_with_sides_mut`], without the addresses.
-    pub fn panes_mut(&mut self) -> impl Iterator<Item = &mut ChartPane> {
-        // Destructured rather than borrowed field by field: the flow pane and
-        // the context stack are two disjoint parts of `self`, and the compiler
-        // only knows that when it is told in one pattern.
-        let Self {
-            flow_pane,
-            time_panes,
-            ..
-        } = self;
-        std::iter::once(flow_pane).chain(time_panes.iter_mut())
-    }
-
-    /// The flow pane's tape.
-    ///
-    /// The flow pane is built with one ([`ChartPane::flow`]) and never gives it
-    /// up; the `Option` on the pane exists so a *time* pane can go without a
-    /// book worker, not because this one can be missing.
-    pub fn tape(&self) -> &OrderflowView {
-        self.flow_pane
-            .orderflow
-            .as_ref()
-            .expect("the flow pane is built with a tape and never drops it")
-    }
-
-    /// See [`Self::tape`].
-    pub fn tape_mut(&mut self) -> &mut OrderflowView {
-        self.flow_pane
-            .orderflow
-            .as_mut()
-            .expect("the flow pane is built with a tape and never drops it")
-    }
-
-    /// Switch which panes the canvas shows (§11).
-    ///
-    /// The first layout that needs the time pane builds it and seeds it from
-    /// the trades the flow pane already holds, so it opens showing the same
-    /// market rather than an empty chart waiting for the next print. Leaving
-    /// a layout only stops drawing its pane: indicators, drawings and bars
-    /// survive, and the pane keeps being fed, so re-showing it never has to
-    /// catch up.
-    ///
-    /// Focus follows what the switch reveals: the pane that just appeared is
-    /// the one the user asked to work on, so the chrome speaks for it without
-    /// a second click. A switch that reveals nothing new (Time + Flow from
-    /// Time + Flow) keeps the focus where it was.
-    pub fn set_layout(&mut self, layout: CanvasLayout) {
-        let previous = self.layout;
-        // A layout chosen from the picker shows the panes its thumbnail drew.
-        // Leaving the column collapsed meant the cell lit, two panes were
-        // promised, and an 8 px rail arrived instead. Before the early return,
-        // because picking the arrangement that is *already* selected is
-        // exactly how a trader asks for the charts they can see promised in a
-        // lit cell — and a return above this line answered that with the rail.
-        self.context_collapsed = false;
-        if layout == previous {
-            return;
-        }
-        self.layout = layout;
-        let wanted = layout
-            .kinds()
-            .iter()
-            .filter(|kind| matches!(kind, PaneKind::Time))
-            .count();
-        // Set, never raised: switching to a three-pane layout and back before
-        // the next frame used to leave the count where the wider layout put
-        // it, and the tab then built a pane no layout had asked for — seeded
-        // from the whole retained tape and fed every trade thereafter.
-        self.pending_context_panes = wanted.saturating_sub(self.time_panes.len());
-        if wanted > self.time_panes.len() {
-            // Seeding replays every retained trade, which on a deep history
-            // holds the render thread long enough to notice. Armed here and
-            // done on the next frame, exactly as a bar-spec change is: the
-            // frame carrying the menu click paints the loading overlay first,
-            // so the wait reads as the chart working rather than the app
-            // hanging.
-            self.loading.begin(LoadingTask::BarRebuild);
-        }
-        self.focus = match layout {
-            CanvasLayout::Single => PaneSide::Flow,
-            CanvasLayout::Time => PaneSide::Time(0),
-            // The split reveals whichever pane the previous layout was not
-            // showing: the time pane coming from Single, the flow pane coming
-            // from Time.
-            CanvasLayout::TimeAndFlow | CanvasLayout::TimeTimeAndFlow => match previous {
-                CanvasLayout::Single => PaneSide::Time(0),
-                CanvasLayout::Time => PaneSide::Flow,
-                CanvasLayout::TimeAndFlow | CanvasLayout::TimeTimeAndFlow => self.focus,
-            },
-        };
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "CANVAS_LAYOUT",
-            layout = ?layout,
-            time_pane_bars = self.time_pane().map(|pane| pane.state.bars().len()),
-            action = if self.pending_context_panes > 0 {
-                "build_time_pane_next_frame"
-            } else {
-                "relayout_canvas"
-            },
-            "canvas layout changed"
-        );
-    }
-
-    /// Build the context panes the last layout change asked for, if any are
-    /// due.
-    ///
-    /// Runs at the top of the frame after the click, so the overlay armed by
-    /// [`Self::set_layout`] has already been painted once. `ids` is the
-    /// window's allocator rather than the tab's: pane ids namespace egui
-    /// interaction state across the whole window, so a tab may not mint its
-    /// own.
-    pub fn apply_pending_layout(
-        &mut self,
-        config: &AppConfig,
-        style: &ChartStyle,
-        ids: &mut PaneIdAllocator,
-    ) {
-        if self.pending_context_panes == 0 {
-            return;
-        }
-        self.pending_context_panes -= 1;
-        // The slot this pane will take is the next free one in the stack.
-        let interval_ms = self
-            .context_opening_intervals_ms
-            .get(self.time_panes.len())
-            .copied()
-            .unwrap_or(self.time_pane_opening_interval_ms);
-        let mut pane = ChartPane::time(ids.alloc(), interval_ms);
-        pane.legend_collapsed = self.time_pane_opening_legend_collapsed;
-        pane.layout = self
-            .context_opening_layouts
-            .get(self.time_panes.len())
-            .copied()
-            .flatten()
-            .map(crate::layouts::LayoutId);
-        pane.seed_from(
-            self.flow_pane.state.trades(),
-            self.flow_pane.state.backfill_trade_count(),
-        );
-        // The pane opens looking like the one it splits away from: a user who
-        // switched the crosshair off is not asking for it back by opening a
-        // second view of the same market. Orientation is part of that look —
-        // an upside-down market does not turn back over by being given a
-        // second view, and the boot's QUANTICK_INVERTED hook fires before
-        // this pane exists at all.
-        // Copying the *switches* rather than a list of field names: an earlier
-        // version cloned `hidden_layers` alone, which left the footprint
-        // behind — a per-pane field of its own — so the split opened with the
-        // ladder on in the flow pane and off in the time pane, contradicting
-        // the paragraph above and darkening the toolbar's footprint lamp the
-        // moment the trader clicked into the left chart. `apply_layer_states`
-        // drops whatever this pane does not draw, which is the whole of what
-        // §11 asks for, and it covers `hidden_layers` in passing.
-        //
-        // Every layer but one. `Drawings` resolves to `DrawingStore`, whose
-        // setter records an undo entry — right for a click, wrong for a pane
-        // being born: seeded through it, a time pane holding zero objects
-        // opens with a non-empty history, and the trader's first Ctrl+Z there
-        // un-hides drawings rather than doing nothing. It is seeded through
-        // the store's own opening setter instead.
-        let mut states = self.flow_pane.layer_states(style);
-        states.remove(&ChartLayer::Drawings);
-        pane.apply_layer_states(&states);
-        pane.drawings
-            .open_all_hidden(self.flow_pane.drawings.all_hidden());
-        pane.price_view
-            .set_inverted(self.flow_pane.price_view.is_inverted());
-        self.time_panes.push(pane);
-        // One pane per frame, for the reason the first one waits a frame at
-        // all: seeding replays every retained trade, and building three at
-        // once would hold the render thread for three times as long. The
-        // overlay stays up until the last one lands.
-        if self.pending_context_panes == 0 {
-            self.loading.end(LoadingTask::BarRebuild);
-        }
-        // The pane exists now, so there is something for a prefix to go in
-        // front of. A base already held (a layout toggled off and on) is
-        // folded rather than re-fetched; otherwise this is the first moment
-        // asking for one means anything.
-        if self.ohlcv_base.is_some() {
-            self.refold_history_prefix();
-        } else {
-            self.request_ohlcv_history(config);
-        }
-    }
-
     /// The display name of the currently selected feed, or its id as a
     /// fallback.
     pub fn feed_display_name<'a>(&'a self, config: &'a AppConfig) -> &'a str {
@@ -1425,232 +861,6 @@ impl Tab {
             return None;
         }
         self.forced_latency.or(*self.feed_latency.borrow())
-    }
-
-    /// The layout each pane opens on, from a saved workspace: the flow pane's
-    /// now, each context pane's when it is built.
-    ///
-    /// `context` is the wire form — [`crate::ui_state::LAYOUT_UNRECORDED`]
-    /// for a slot the file did not state — and it is read back into an
-    /// `Option` here, at the one place the file's shape meets the pane's.
-    ///
-    /// **Ordering.** This names what a pane *will* show; it does not move a
-    /// pane that is already showing something. A caller that reaches a tab
-    /// whose stack is standing must follow with
-    /// `QuantickApp::reload_layouts`, which clears every pane's set and seeds
-    /// it again from the layout named here — otherwise a seeded pane keeps
-    /// the old layout's indicators, drawings and header label under the new
-    /// layout's id, and the next edit is written into the wrong entries.
-    /// Both callers do (`restore_workspace` and the bundle import, each
-    /// through `reload_cockpit_stores`).
-    pub fn set_opening_layouts(&mut self, flow: Option<u64>, context: &[u64]) {
-        self.flow_pane.layout = flow.map(crate::layouts::LayoutId);
-        self.context_opening_layouts = context
-            .iter()
-            .map(|id| (*id != crate::ui_state::LAYOUT_UNRECORDED).then_some(*id))
-            .take(MAX_CONTEXT_PANES)
-            .collect();
-        // A context pane already built takes its name now: an import lands on
-        // a tab whose stack is standing, and the stash alone would only reach
-        // the panes still to come.
-        for (slot, pane) in self.time_panes.iter_mut().enumerate() {
-            if let Some(id) = context
-                .get(slot)
-                .copied()
-                .filter(|id| *id != crate::ui_state::LAYOUT_UNRECORDED)
-            {
-                pane.layout = Some(crate::layouts::LayoutId(id));
-            }
-        }
-    }
-
-    /// Name the layout a context pane will open on when the stack builds it.
-    ///
-    /// Only for a pane that is not standing yet — a canvas change lands its
-    /// panes a frame after the layout that asked for them. A pane already on
-    /// the canvas is moved by `QuantickApp::switch_pane_layout`, which
-    /// carries its indicators and its drawings across too; writing
-    /// `pane.layout` under a standing pane would leave the field disagreeing
-    /// with what that chart is showing.
-    pub fn set_opening_layout(&mut self, side: PaneSide, id: crate::layouts::LayoutId) {
-        // The flow pane is built with the tab and is never pending, so the
-        // only address that can be waiting is a context slot.
-        let PaneSide::Time(slot) = side else {
-            debug_assert!(false, "the flow pane is never waiting to be built");
-            return;
-        };
-        if slot >= MAX_CONTEXT_PANES {
-            return;
-        }
-        if self.context_opening_layouts.len() <= slot {
-            self.context_opening_layouts.resize(slot + 1, None);
-        }
-        self.context_opening_layouts[slot] = Some(id.0);
-    }
-
-    /// Put this tab's canvas back the way a saved workspace recorded it: the
-    /// layout, the divider, the focused pane, and the interval the time pane
-    /// opens on.
-    ///
-    /// One method rather than four public fields, because the order matters
-    /// and only the tab knows it. The opening interval has to be set *before*
-    /// [`Self::set_layout`] arms the time pane, or the pane is built on the
-    /// header default and the saved interval lands one frame too late. The
-    /// focus is applied *after*, because `set_layout` moves it to whatever the
-    /// switch reveals — right for a menu click, wrong for a restore, where the
-    /// saved focus is the answer.
-    ///
-    /// Startup-scoped, like [`Self::apply_feed_declared_layout`]: the only
-    /// caller is the app restoring a workspace into a tab it has just opened.
-    pub fn restore_canvas(
-        &mut self,
-        layout: CanvasLayout,
-        split_fraction: Option<f32>,
-        context_collapsed: bool,
-        focus: Option<PaneSide>,
-        context_intervals_ms: &[i64],
-        legends: LegendFold,
-    ) {
-        // The top chart's interval is also the one every slot past the list
-        // opens on, which is what a one-chart file has always meant.
-        if let Some(ms) = context_intervals_ms.first() {
-            self.time_pane_opening_interval_ms = *ms;
-        }
-        self.context_opening_intervals_ms = context_intervals_ms
-            .iter()
-            .copied()
-            .take(MAX_CONTEXT_PANES)
-            .collect();
-        self.set_layout(layout);
-        // *After* `set_layout`, for the reason the focus below is: a switch
-        // opens the column it just revealed, which is right for a menu click
-        // and wrong for a restore, where the saved state is the answer.
-        // Assigned before, a workspace saved with its charts put away reopened
-        // with them out — and the next `capture_arrangement` wrote that over
-        // the trader's choice.
-        self.context_collapsed = context_collapsed;
-        if let Some(fraction) = split_fraction {
-            self.split_fraction = clamp_pane_fraction(fraction);
-        }
-        if let Some(side) = focus {
-            self.focus = side;
-        }
-        self.flow_pane.legend_collapsed = legends.flow;
-        // The time pane may not exist yet. `set_layout` only *arms* it
-        // (`pending_time_pane`); `apply_pending_layout` builds it a frame
-        // later, which is the very reason the opening interval above is
-        // stashed rather than assigned. Writing the fold here alone would
-        // write it to `None` and the pane would open expanded — and the next
-        // `capture_arrangement` would then persist that `false` over the
-        // trader's choice.
-        self.time_pane_opening_legend_collapsed = legends.time;
-        if let Some(time) = self.time_pane_mut() {
-            time.legend_collapsed = legends.time;
-        }
-    }
-
-    /// Let every pane's selectors settle, then mirror the result onto the
-    /// rebuild indicator: it is up while *any* pane has a rebuild pending.
-    pub fn apply_spec_changes(&mut self) {
-        // Every pane, by address. Settling "the flow pane and the time pane"
-        // left the second stacked chart's selector armed for ever: its header
-        // chip lit, its interval changed, and its bars never rebuilt.
-        for pane in 0..self.pane_count() {
-            self.apply_spec_change_at(pane);
-        }
-        let rebuilding = self
-            .panes()
-            .any(|(pane, _side)| pane.spec.pending.is_some());
-        self.loading.set_active(LoadingTask::BarRebuild, rebuilding);
-    }
-
-    /// Apply one pane's bar-type/parameter change, a frame after its selectors
-    /// settle.
-    ///
-    /// Switching the spec replays every retained trade synchronously, which
-    /// can hold this thread long enough to notice on a deep history. Deferring
-    /// the rebuild by one frame lets the frame that carries the change paint
-    /// the loading overlay first, so the wait reads as the chart working
-    /// rather than the app hanging. A selector still moving (a dragged
-    /// parameter) keeps pushing the pending spec forward, which also debounces
-    /// the rebuild to one per gesture.
-    ///
-    /// The two panes run this independently: the toolbar's BARS group governs
-    /// the focused pane and the time pane's own header governs the time pane
-    /// (§11), so a change to one pane must not rebuild the chart beside it.
-    fn apply_spec_change_at(&mut self, index: PaneIndex) {
-        let Some(desired) = self.pane_at(index).map(ChartPane::current_spec) else {
-            return;
-        };
-        let Some(pane) = self.pane_at_mut(index) else {
-            return;
-        };
-        if desired == *pane.state.spec() {
-            // Selection and chart agree — nothing is pending any more (a feed
-            // switch or reset may have rebuilt the state under a pending spec).
-            pane.spec.pending = None;
-            return;
-        }
-        match pane.spec.pending.take() {
-            // The frame that changed the selector: arm the indicator, paint.
-            None => pane.spec.pending = Some(desired),
-            // Still moving: wait for the selector to settle for a frame.
-            Some(pending) if pending != desired => pane.spec.pending = Some(desired),
-            // Settled since last frame: do the rebuild.
-            Some(_) => {
-                // Where the user is looking, in market time — the one thing a
-                // rebuild preserves. The new series cuts the same trades into
-                // a different number of bars, so the old right-edge *index*
-                // may not exist in it at all: keeping it would leave the
-                // window past the end of the data, drawing nothing.
-                let anchor = pane.right_edge_time();
-                // The series the drawings are still anchored to, captured
-                // before it is replaced: their bar indices are meaningless in
-                // the new cut and have to be re-derived from the market time
-                // each anchor carries.
-                let old_slots = pane.slots();
-                pane.set_spec(desired);
-                // The venue prefix folds to the new interval before the view
-                // is reanchored: the market time the user was looking at has
-                // to resolve against the series they will be looking at.
-                // Either pane: `bars → time` on the flow pane is exactly the
-                // spec change this refold exists for (audit S1).
-                //
-                // Installing a prefix rebuilds the indicators over the whole
-                // composed series, so the plain rebuild is only sent when the
-                // refold did not — two rebuilds of ~130k bars per settled drag
-                // frame, with no coalescing in the worker, is the cost of
-                // sending both.
-                let refolded = self.refold_history_prefix();
-                if !refolded && let Some(pane) = self.pane_at_mut(index) {
-                    pane.send_indicator_rebuild();
-                }
-                let Some(pane) = self.pane_at_mut(index) else {
-                    return;
-                };
-                let slot = anchor.and_then(|ms| pane.slot_at_time(ms));
-                let slots = pane.slots();
-                pane.viewport.reanchor(slot, slots);
-                // The marks follow the view: same market time, this pane's
-                // new bar space. Nothing is lost, so there is nothing to
-                // announce.
-                pane.reanchor_drawings(old_slots);
-                // The strategies do not follow: the body average that
-                // defines a force bar means something else under another
-                // bar spec, so the instances disarm and say why. The tape
-                // itself continues, so any pending bot entry is swept here
-                // and now — through the same funnel manual orders use.
-                let cleanup = pane
-                    .strategies
-                    .anchors
-                    .disarm_all(quantick_strategy::DisarmReason::BarSpecChanged);
-                let _ = pane.take_strategy_bars();
-                for command in cleanup {
-                    let _ = self.paper.account_mut().apply_strategy_command(command);
-                }
-                self.drop_overlay_gestures();
-            }
-        }
     }
 }
